mldsa: a crafted message length must not halt the chain
The batch-verify path narrowed an attacker-chosen 256-bit length word to int.
Any value with bit 63 set becomes NEGATIVE, and a negative length SATISFIES
`offset+msgLen > len(input)` -- the sum is smaller than the length, not larger --
so the bounds check waves it through and the slice expression panics on a
negative bound:

    slice bounds out of range [:-9223372036854770511]

The panic is unrecovered and occurs inside block verification, so one
transaction to 0x0000000000000000000000000000000000012202 halts every validator
that verifies the block carrying it. The precompile is genesis-active on
mainnet -- eth_getCode returns 0x01 -- so nothing gates it.

Fixed the way the single-verify path in this same file already does it: keep the
length in uint64 and compare in uint64, so a huge value cannot wrap into
something that looks in-bounds. After that check msgLen fits an int by
construction, bounded by len(input)-offset.

The regression test asserts NotPanics, because rejecting is not the interesting
half -- surviving is. Reverting the fix reproduces the exact panic above.

+++ mldsa/contract.go
@@ -304,7 +304,14 @@ func (p *mldsaVerifyPrecompile) runBatchVerify(input []byte, suppliedGas, gasCos
 		if offset+MessageLenSize > len(input) {
 			return nil, remainingGas, fmt.Errorf("%w: entry %d truncated at msglen", ErrInvalidInputLength, i)
 		}
-		msgLen := int(readUint256(input[offset : offset+MessageLenSize]))
+		// Stay in uint64, as the single-verify path above does. Narrowing an
+		// attacker-chosen 256-bit word to int makes any value with bit 63 set
+		// NEGATIVE, and a negative length passes `offset+msgLen > len(input)`
+		// because the sum is smaller than the length, not larger. The slice
+		// expression below then panics on a negative bound, unrecovered, inside
+		// block verification -- so one transaction halts every validator that
+		// verifies the block carrying it.
+		msgLen := readUint256(input[offset : offset+MessageLenSize])
 		offset += MessageLenSize
 
 		// signature
@@ -314,12 +321,14 @@ func (p *mldsaVerifyPrecompile) runBatchVerify(input []byte, suppliedGas, gasCos
 		signatures[i] = input[offset : offset+sigSize]
 		offset += sigSize
 
-		// message
-		if offset+msgLen > len(input) {
+		// message. Compared in uint64 so a huge length cannot wrap into a value
+		// that looks in-bounds; after this passes, msgLen fits an int by
+		// construction because it is bounded by len(input)-offset.
+		if uint64(offset)+msgLen > uint64(len(input)) {
 			return nil, remainingGas, fmt.Errorf("%w: entry %d truncated at message", ErrInvalidInputLength, i)
 		}
-		messages[i] = input[offset : offset+msgLen]
-		offset += msgLen
+		messages[i] = input[offset : offset+int(msgLen)]
+		offset += int(msgLen)
 	}
 
 	if offset != len(input) {
